refacotr:suggestion컴포넌트 분리, 훅분리
